Add native support for OIDC (#13)
Use the implementation and configuration from @pulumi/actions-helpers.

Along with the recent changes to pull configuration from environment variables, this makes it straightforward to configure the ESC Action once for an entire workflow using environment variables, which reduces the minimal per-job boilerplate to

    uses: pulumi/esc-action@...

or, for use cases that require the step's outputs,

    id: esc-secrets
    uses: pulumi/esc-action@...

This also required replacing the `ncc` bundler with `rollup`, as I couldn't get the former to work. The latter appears to work nicely (it is what the [official GH TS template](https://github.com/actions/typescript-action/tree/main) uses).