Support key remapping (#15)
If the 'export-environment-variables' input is present and is neither true nor false, interpret it as a list of [FOO=]BAR key-value pairs, where FOO is the name of the envvar to set and BAR is the name of the variable to use as the value. If FOO is omitted, BAR is also used as the name of the envvar to set. If BAR is '*', then all unmapped variables are implicitly mapped to themselves.

For example, the value 'GITHUB_TOKEN=PULUMI_BOT_TOKEN,AWS_KEY_ID, AWS_SECRET_KEY,AWS_SESSION_TOKEN' will export this environment:

    GITHUB_TOKEN=PULUMI_BOT_TOKEN
    AWS_KEY_ID=AWS_KEY_ID
    AWS_SECRET_KEY=AWS_SECRET_KEY
    AWS_SESSION_TOKEN=AWS_SESSION_TOKEN

If the source ESC env also contained other environment variables, they would not be exported. All non-mapped variables can be exported with the identity mapping by including '*' as a key. For example, 'GITHUB_TOKEN=PULUMI_BOT_TOKEN,*` would also export the environment above assuming no other envvars exist in the ESC environment.

Remapping applies to action outputs as well as environment variables.

If `export-environment-variables` is `true`, then `keys` is treated as the source of the mappings to maintain compatibility with existing workflows.