Fix a bug in multiply-mapped envvars (#22)
In the case where a source environment variable was mapped to multiple
destinations, the ESC action would only use the last mapping. These
changes fix the behavior to use all of the listed mappings.

There is a bit of subtlety with respect to the wildcard mapping: if a
variable is otherwise mapped, the wildcard mapping will not add an
identity mapping for that variable. For example, `FOO=BAR,*` will
not add an identity mapping for `BAR`.