Don't trim values and use a more unique delimiter (#32)
The action currently trims values but this can remove meaningful whitespace. For example, a value containing a PEM-encoded private key _must_ include a trailing newline (RFC 7468). Trimming a value containing a PEM key renders it invalid.

Instead we should return the value as-is, including any whitespace the user specified with it.

This also fixes an issue where if "EOF" appeared anywhere in the value it would cause the value to not be set in the environment because GitHub's parsing doesn't seem to handle that correctly (I'm working on a repro to confirm). Instead we use a more distinct "PULUMIESCEOF" delimiter which is practically guaranteed to never appear in the wild.

Refs https://github.com/pulumi/ci-mgmt/issues/1481